View data + callback + fixes
Fixed the view page! Relations, dropdown, multiselect... all that stuff now is shown as the list page!
Finally added the callback_read_field() that wasn't finished since long time.
If callback_read_field() is not set and callback_column is set, the last one is used for the read page (if you've changed that is because you want to see it, right?)
That was tough to fix but so useful to everyone!